# Optimize parsing 19 digit longs