Parsing 19 digit long values is significantly slower than parsing 18 digit long values as we are hitting a slow path that does a string allocation.

19 digit long values are quite common when a random number generator is used to generate long values.

We have to pay attention that we only use this fast path if the value is a valid long. This can be achieved by calling `#inLongRange` first.

In the following synthetic microbenchmark a throughput increase of about 33% with a drastic reduction in the allocation rate can be achieved.

```java

@OutputTimeUnit(TimeUnit.SECONDS)
public class JacksonStdLongReadVanilla
{

    private static final JsonFactory JSON_FACTORY = new JsonFactory();

    private static final String JSON;

    static {
        StringBuilder json = new StringBuilder();
        Random random = new Random(0L);
        json.append('[');
        for (int i = 0; i < 1_000; i++) {
            if (i > 0) {
                json.append(',');
            }
            json.append(random.nextLong());
        }
        json.append(']');
        JSON = json.toString();
    }

    @Benchmark
    public void parseLong(Blackhole blackhole) throws IOException {
        try (JsonParser parser = JSON_FACTORY.createParser(JSON)) {
            parser.nextToken();

            JsonToken nextToken = parser.nextToken();
            while (nextToken == JsonToken.VALUE_NUMBER_INT) {
                blackhole.consume(parser.getLongValue());
                nextToken = parser.nextToken();
            }
        }
    }

}

```

If we have a look at the allocations before we see that it is dominated by `byte[]` allocations for `String` allocations

![allocation-site-before](https://user-images.githubusercontent.com/471021/207910102-4dd41f72-73b0-4aaf-a4f6-541f68e25e53.png)

Resulting in many gcs

![gcs-before](https://user-images.githubusercontent.com/471021/207910097-7642ddad-33ec-4933-9263-faa80c8c6e21.png)

With the changes in this PR the  `byte[]` and `String` allocations are gone

![allocation-site-after](https://user-images.githubusercontent.com/471021/207910103-6dedc3c3-3df7-4d88-96cb-4026c61ae4e9.png)

Resulting in no more gcs

![gcs-after](https://user-images.githubusercontent.com/471021/207910100-76bb889e-551a-4536-8fc9-0cdb95cdfc0f.png)