@pjfanning 2022-12-15
Another thing that I'm concerned with is that with a custom long parser, we would need to have test coverage for a wide range of possible values.

The FastDoubleParser that we allow use of is
* optional (and not the default)
* FastDoubleParser has good test coverage
* it is admittedly still quite experimental and users who opt in to using it will need to do so with caution