@marschall 2022-12-16
> Another thing that I'm concerned with is that with a custom long parser, we would need to have test coverage for a wide range of possible values.
> 

`NumberInput` is already full of custom number parsing code, I refer you to the various `#parseInt` methods. There is also the custom `BigDecimalParser`. But I can extend the tests in `NumberParsingTest`.