@marschall 2022-12-20
> Although `jackson-benchmarks` is useful for comparing performance of various Jackson backends over relatively simple POJO type, it's not as good for micro-benchmarks for aspects like this. So I think I'd go with https://github.com/pjfanning/jackson-number-parse-bench.

There you go https://github.com/pjfanning/jackson-number-parse-bench/pull/2

I had to change `build.gradle`

```groovy
repositories {
    mavenLocal()
}

jmh {
    includes = ['org.example.jackson.bench.LongParserBench']
    profilers = ['gc'] // or whatever you like best like jfr or async
}
```

On a Ryzen 7 PRO 5750GE we go from

```
Benchmark                                                    Mode  Cnt      Score      Error   Units
LongParserBench.parseLong                                   thrpt   25  21860.758 ± 1076.882   ops/s
LongParserBench.parseLong:·gc.alloc.rate                    thrpt   25   1139.800 ±   56.144  MB/sec
LongParserBench.parseLong:·gc.alloc.rate.norm               thrpt   25  57410.238 ±    0.043    B/op
LongParserBench.parseLong:·gc.churn.G1_Eden_Space           thrpt   25   1136.527 ±   59.318  MB/sec
LongParserBench.parseLong:·gc.churn.G1_Eden_Space.norm      thrpt   25  57240.652 ±  766.795    B/op
LongParserBench.parseLong:·gc.churn.G1_Survivor_Space       thrpt   25      0.006 ±    0.002  MB/sec
LongParserBench.parseLong:·gc.churn.G1_Survivor_Space.norm  thrpt   25      0.303 ±    0.079    B/op
LongParserBench.parseLong:·gc.count                         thrpt   25    504.000             counts
LongParserBench.parseLong:·gc.time                          thrpt   25    334.000                 ms
```

to 
```
Benchmark                                                    Mode  Cnt      Score      Error   Units
LongParserBench.parseLong                                   thrpt   25  31230.896 ± 1532.563   ops/s
LongParserBench.parseLong:·gc.alloc.rate                    thrpt   25     14.523 ±    0.713  MB/sec
LongParserBench.parseLong:·gc.alloc.rate.norm               thrpt   25    512.025 ±    0.029    B/op
LongParserBench.parseLong:·gc.churn.G1_Eden_Space           thrpt   25     15.785 ±   19.352  MB/sec
LongParserBench.parseLong:·gc.churn.G1_Eden_Space.norm      thrpt   25    544.799 ±  668.200    B/op
LongParserBench.parseLong:·gc.churn.G1_Survivor_Space       thrpt   25     ≈ 10⁻⁴             MB/sec
LongParserBench.parseLong:·gc.churn.G1_Survivor_Space.norm  thrpt   25      0.005 ±    0.011    B/op
LongParserBench.parseLong:·gc.count                         thrpt   25      7.000             counts
LongParserBench.parseLong:·gc.time                          thrpt   25     10.000                 ms
```

We go from 21860.758 ops/s to 31230.896 ops/s.
We go from an allocation rate of 1139 MB/sec to 14 MB/sec.
We go from 504 gcs taking 334 ms to 7 gc taking 10 ms.