@marschall 2022-12-21
> > Could NumberInput.parseLong be changed to call parseLong19 where appropriate - it already uses parseInt for small nums. Would it be simpler just to have parseLong parse all numbers the new way (if the performance was not worse)?
> 
> Right now method is quite particular expecting length of 10 - 18 characters and there'd be one more check (and branch). Not sure it would affect optimizer that much; I don't have strong opinion where to modify it and add checks or keep separate method. Separate method would do well to follow the same pattern tho; decoding 2 ints, using `int` multiplications is faster than doing 19 `long` multiplications. I remember benchmarking that difference years ago and it was non-trivial.

I had a look at this, it would require a bit more changes. Currently `#parseInt` only handles `int`s with length 9. `int`s with length 10 are parsed as `long` and then converted to `int`. Therefore we can't call `#parseInt` twice to parse `long` of length 19, only `long` of length 18.

If we we wanted to change this we would have to:
- Introduce `NumberInput#inIntRange`
- Change `ParserBase#_parseIntValue()` and `ParserBase#_parseNumericValue(int)` to call `NumberInput#inIntRange` for length 10 and based on this either call `TextBuffer#contentsAsInt(boolean)` or `TextBuffer#contentsAsLong(boolean)`.
- Change `#parseInt` to handle a length of up to 10