@cowtowncoder 2022-12-21
I think we are ok for now; if anyone really wants to optimize things further, a new PR would be fine.
There's fine line between handling 10/19 lengths optimally vs handling shorter ones -- for most use cases I suspect it is more important to have close(r) to optimal parsing for shorter numbers than longer. That is: adding more checks (-> more branching in code) for different cases slows down existing cases.